refactor: Split components out of Debug page
